+++ APPLICATION/DTOs/Redemptions/RedemptionDetailDto.cs
@@ -7,7 +7,7 @@ public class RedemptionDetailDto
     public Guid Id { get; set; }
     
     // Redemption info
-    public int PointsSpent { get; set; }
+    public decimal PointsSpent { get; set; }
     public int Quantity { get; set; }
     public RedemptionStatus Status { get; set; }
     public string? AdminNotes { get; set; }
@@ -24,9 +24,9 @@ public class RedemptionDetailDto
     public string UserName { get; set; } = string.Empty;
     public string UserEmail { get; set; } = string.Empty;
     public string UserAvatarUrl { get; set; } = string.Empty;
-    public int UserCurrentBalance { get; set; }
-    public int UserTotalEarned { get; set; }
-    public int UserTotalRedeemed { get; set; }
+    public decimal UserCurrentBalance { get; set; }
+    public decimal UserTotalEarned { get; set; }
+    public decimal UserTotalRedeemed { get; set; }
     
     // Product snapshot
     public Guid ProductId { get; set; }
@@ -34,6 +34,6 @@ public class RedemptionDetailDto
     public string ProductDescription { get; set; } = string.Empty;
     public string ProductCategory { get; set; } = string.Empty;
     public string ProductImageUrl { get; set; } = string.Empty;
-    public int ProductPointsPerUnit { get; set; }
-    public int ProductTotalPoints { get; set; }
+    public decimal ProductPointsPerUnit { get; set; }
+    public decimal ProductTotalPoints { get; set; }
 }
